Use float for draw angle, instead of int.

+++ src/battle/debris.h
@@ -26,7 +26,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 #define MAX_DEBRIS_TEXTURES		6
 
 extern float mod(float n, float x);
-extern void blitRotated(SDL_Texture *texture, int x, int y, int angle);
+extern void blitRotated(SDL_Texture *texture, int x, int y, float angle);
 extern int collision(int x1, int y1, int w1, int h1, int x2, int y2, int w2, int h2);
 extern SDL_Texture *getTexture(char *filename);
 extern void addDebrisFire(int x, int y);
